--- lib/servactory/inputs/define_input_conflict.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Servactory
-  module Inputs
-    class DefineInputConflict
-      attr_reader :content
-
-      def initialize(content:)
-        @content = content
-      end
-    end
-  end
-end
