+++ lib/servactory/maintenance/attributes/define_conflict.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Servactory
+  module Maintenance
+    module Attributes
+      class DefineConflict
+        attr_reader :content
+
+        def initialize(content:)
+          @content = content
+        end
+      end
+    end
+  end
+end
