--- sig/lib/servactory/inputs/define_input_conflict.rbs
@@ -1,9 +1,0 @@
-module Servactory
-  module Inputs
-    class DefineInputConflict
-      attr_reader content: Proc
-
-      def initialize: (content: Proc) -> void
-    end
-  end
-end
