--- sig/lib/servactory/inputs/define_input_method.rbs
@@ -1,11 +1,0 @@
-module Servactory
-  module Inputs
-    class DefineInputMethod
-      attr_reader name: Symbol
-
-      attr_reader content: Proc
-
-      def initialize: (name: Symbol, content: Proc) -> void
-    end
-  end
-end
