+++ sig/lib/servactory/maintenance/attributes/define_conflict.rbs
@@ -1,0 +1,11 @@
+module Servactory
+  module Maintenance
+    module Attributes
+      class DefineConflict
+        attr_reader content: Proc
+
+        def initialize: (content: Proc) -> void
+      end
+    end
+  end
+end
